Redux Toolkit을 사용하는 테스트로 수정
redux-thunk를 uninstall 했으므로 redux toolkit을 사용하여 미들웨어를
얻어오고 이를 테스트에 적용한다.

그리고 slice에 action과 reducer 테스트를 합친다.
