+++ src/store.js
@@ -1,9 +1,7 @@
-import { createStore, applyMiddleware } from 'redux';
-
-import thunk from 'redux-thunk';
+import { createStore } from 'redux';
 
 import reducer from './slice';
 
-const store = createStore(reducer, applyMiddleware(thunk));
+const store = createStore(reducer);
 
 export default store;
